backend/pyproject.toml
[project]
name = "agentos-backend"
version = "0.1.0"
description = "AgentOS backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pyyaml>=6.0.2",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "openai>=1.55.0",
  "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
